[#816] Pulled pure dashboard functions out into separate utils namesapce
